Trying to think of new ways to store team info
The problem with just defining a table to store all the data is that each year, the game is different. As such, the data to be collected is also different. I'm trying to think about how to get around that problem. The best solution I've come up with so far, is a MongoDB database to hold the data. There would be one collection actually holding the data, and there would be a second collection defining the data layout for each year.