Update data if Schema changed
If a key is renamed or deleted, update the stored data to match.
If a key has its display name or type changed, the data will not be
updated. The user will be prompted to update the data to conform to the
new type next time the user tries to edit that team's data